* Added compilation flags for extra warnings (C++) * Cleaned code

+++ engine/include/core/DateTime.h
@@ -19,7 +19,7 @@ namespace openfluid { namespace core {
 /**
   Type for raw time (seconds since 4713BC)
 */
-typedef long long rawtime_t;
+typedef unsigned long rawtime_t;
 
 /**
   \brief Class for management of date and time information.
